+++ koupleless-base-plugin/src/main/java/com/alipay/sofa/koupleless/plugin/spring/BizApplicationContextInitializer.java
@@ -18,7 +18,7 @@ package com.alipay.sofa.koupleless.plugin.spring;
 
 import com.alipay.sofa.koupleless.common.BizRuntimeContext;
 import com.alipay.sofa.koupleless.common.BizRuntimeContextRegistry;
-import com.alipay.sofa.koupleless.common.model.SpringBizApplicationContext;
+import com.alipay.sofa.koupleless.common.model.SpringApplicationContextHolder;
 import org.springframework.context.ApplicationContextInitializer;
 import org.springframework.context.ConfigurableApplicationContext;
 
@@ -31,7 +31,7 @@ public class BizApplicationContextInitializer implements
         BizRuntimeContext bizRuntimeContext = BizRuntimeContextRegistry
             .getBizRuntimeContextByClassLoader(classLoader);
         bizRuntimeContext
-            .setApplicationContext(new SpringBizApplicationContext(applicationContext));
+            .setApplicationContext(new SpringApplicationContextHolder(applicationContext));
         bizRuntimeContext.setRootApplicationContext(applicationContext);
     }
 }
